Add files via upload
I applied various machine learning algorithms to the dataset and compared the performance on each algorithm on how they predicted the target class. The aim of the project was to use create an algorithm that can accurately predict if a patient is Abnormal or Normal based on some biomechanics measurement taken.